Update transactions to passthrough privy without lossy serialization
